Move RenderEdit methods to separate types
