+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/DaggerSpawnEvents.cs
@@ -52,63 +52,4 @@ public sealed class DaggerSpawnEvents : IEventTypeRenderer<DaggerSpawnEventData>
 		EventTypeRendererUtils.NextColumnInputInt16Mat3x3(eventIndex, nameof(DaggerSpawnEventData.Orientation), ref e.Orientation);
 		EventTypeRendererUtils.NextColumnCheckbox(eventIndex, nameof(DaggerSpawnEventData.IsShot), ref e.IsShot, "Shot", "Rapid");
 	}
-
-	public static void RenderEdit(int uniqueId, DaggerSpawnEventData e, EditorReplayModel replay)
-	{
-		const float leftColumnWidth = 120;
-		const float rightColumnWidth = 160;
-		const float tableWidth = leftColumnWidth + rightColumnWidth;
-
-		if (ImGui.BeginChild(Inline.Span($"DaggerSpawnEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
-		{
-			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Type");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputByteEnum(uniqueId, nameof(DaggerSpawnEventData.DaggerType), ref e.DaggerType, EnumUtils.DaggerTypes, _daggerTypeNamesArray);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt(uniqueId, nameof(DaggerSpawnEventData.A), ref e.A);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Shot/Rapid");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.Checkbox(uniqueId, nameof(DaggerSpawnEventData.IsShot), ref e.IsShot, "Shot", "Rapid");
-
-				ImGui.EndTable();
-			}
-
-			ImGui.SameLine();
-
-			if (ImGui.BeginTable("Right", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("RightText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("RightInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Position");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt16Vec3(uniqueId, nameof(DaggerSpawnEventData.Position), ref e.Position);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Orientation");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt16Mat3x3Square(uniqueId, nameof(DaggerSpawnEventData.Orientation), ref e.Orientation);
-
-				ImGui.EndTable();
-			}
-		}
-
-		ImGui.EndChild();
-	}
 }
