+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/InitialInputsEvents.cs
@@ -52,8 +52,4 @@ public sealed class InitialInputsEvents : IEventTypeRenderer<InitialInputsEventD
 		EventTypeRendererUtils.NextColumnInputShort(eventIndex, nameof(InitialInputsEventData.MouseY), ref e.MouseY);
 		EventTypeRendererUtils.NextColumnInputFloat(eventIndex, nameof(InitialInputsEventData.LookSpeed), ref e.LookSpeed, "%.2f");
 	}
-
-	public static void RenderEdit(int uniqueId, InitialInputsEventData e, EditorReplayModel replay)
-	{
-	}
 }
