+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/InputsEvents.cs
@@ -50,8 +50,4 @@ public sealed class InputsEvents : IEventTypeRenderer<InputsEventData>
 		EventTypeRendererUtils.NextColumnInputShort(eventIndex, nameof(InputsEventData.MouseX), ref e.MouseX);
 		EventTypeRendererUtils.NextColumnInputShort(eventIndex, nameof(InputsEventData.MouseY), ref e.MouseY);
 	}
-
-	public static void RenderEdit(int uniqueId, InputsEventData e, EditorReplayModel replay)
-	{
-	}
 }
