+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/SpiderSpawnEvents.cs
@@ -39,41 +39,4 @@ public sealed class SpiderSpawnEvents : IEventTypeRenderer<SpiderSpawnEventData>
 		EventTypeRendererUtils.NextColumnInputInt(eventIndex, nameof(SpiderSpawnEventData.A), ref e.A);
 		EventTypeRendererUtils.NextColumnInputVector3(eventIndex, nameof(SpiderSpawnEventData.Position), ref e.Position, "%.2f");
 	}
-
-	public static void RenderEdit(int uniqueId, SpiderSpawnEventData e, EditorReplayModel replay)
-	{
-		const float leftColumnWidth = 120;
-		const float rightColumnWidth = 160;
-		const float tableWidth = leftColumnWidth + rightColumnWidth;
-
-		if (ImGui.BeginChild(Inline.Span($"SpiderSpawnEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
-		{
-			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Type");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputByteEnum(uniqueId, nameof(SpiderSpawnEventData.SpiderType), ref e.SpiderType, EnumUtils.SpiderTypes, _spiderTypeNamesArray);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Position");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputVector3(uniqueId, nameof(SpiderSpawnEventData.Position), ref e.Position, "%.2f");
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt(uniqueId, nameof(SpiderSpawnEventData.A), ref e.A);
-
-				ImGui.EndTable();
-			}
-		}
-
-		ImGui.EndChild();
-	}
 }
