+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/SquidSpawnEvents.cs
@@ -43,63 +43,4 @@ public sealed class SquidSpawnEvents : IEventTypeRenderer<SquidSpawnEventData>
 		EventTypeRendererUtils.NextColumnInputVector3(eventIndex, nameof(SquidSpawnEventData.Direction), ref e.Direction, "%.2f");
 		EventTypeRendererUtils.NextColumnInputFloat(eventIndex, nameof(SquidSpawnEventData.RotationInRadians), ref e.RotationInRadians, "%.2f");
 	}
-
-	public static void RenderEdit(int uniqueId, SquidSpawnEventData e, EditorReplayModel replay)
-	{
-		const float leftColumnWidth = 120;
-		const float rightColumnWidth = 160;
-		const float tableWidth = leftColumnWidth + rightColumnWidth;
-
-		if (ImGui.BeginChild(Inline.Span($"SquidSpawnEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
-		{
-			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Type");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputByteEnum(uniqueId, nameof(SquidSpawnEventData.SquidType), ref e.SquidType, EnumUtils.SquidTypes, _squidTypeNamesArray);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt(uniqueId, nameof(SquidSpawnEventData.A), ref e.A);
-
-				ImGui.EndTable();
-			}
-
-			ImGui.SameLine();
-
-			if (ImGui.BeginTable("Right", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("RightText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("RightInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Position");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputVector3(uniqueId, nameof(SquidSpawnEventData.Position), ref e.Position, "%.2f");
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Direction");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputVector3(uniqueId, nameof(SquidSpawnEventData.Direction), ref e.Direction, "%.2f");
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Rotation");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputFloat(uniqueId, nameof(SquidSpawnEventData.RotationInRadians), ref e.RotationInRadians, "%.2f");
-
-				ImGui.EndTable();
-			}
-		}
-
-		ImGui.EndChild();
-	}
 }
