+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/ThornSpawnEvents.cs
@@ -36,41 +36,4 @@ public sealed class ThornSpawnEvents : IEventTypeRenderer<ThornSpawnEventData>
 		EventTypeRendererUtils.NextColumnInputVector3(eventIndex, nameof(ThornSpawnEventData.Position), ref e.Position, "%.2f");
 		EventTypeRendererUtils.NextColumnInputFloat(eventIndex, nameof(ThornSpawnEventData.RotationInRadians), ref e.RotationInRadians, "%.2f");
 	}
-
-	public static void RenderEdit(int uniqueId, ThornSpawnEventData e, EditorReplayModel replay)
-	{
-		const float leftColumnWidth = 120;
-		const float rightColumnWidth = 160;
-		const float tableWidth = leftColumnWidth + rightColumnWidth;
-
-		if (ImGui.BeginChild(Inline.Span($"ThornSpawnEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
-		{
-			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Position");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputVector3(uniqueId, nameof(ThornSpawnEventData.Position), ref e.Position, "%.2f");
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Rotation");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputFloat(uniqueId, nameof(ThornSpawnEventData.RotationInRadians), ref e.RotationInRadians, "%.2f");
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt(uniqueId, nameof(ThornSpawnEventData.A), ref e.A);
-
-				ImGui.EndTable();
-			}
-		}
-
-		ImGui.EndChild();
-	}
 }
