+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Events/EventTypes/TransmuteEvents.cs
@@ -38,53 +38,4 @@ public sealed class TransmuteEvents : IEventTypeRenderer<TransmuteEventData>
 		EventTypeRendererUtils.NextColumnInputInt16Vec3(eventIndex, nameof(TransmuteEventData.C), ref e.C);
 		EventTypeRendererUtils.NextColumnInputInt16Vec3(eventIndex, nameof(TransmuteEventData.D), ref e.D);
 	}
-
-	public static void RenderEdit(int uniqueId, TransmuteEventData e, EditorReplayModel replay)
-	{
-		const float leftColumnWidth = 120;
-		const float rightColumnWidth = 160;
-		const float tableWidth = leftColumnWidth + rightColumnWidth;
-
-		if (ImGui.BeginChild(Inline.Span($"TransmuteEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
-		{
-			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
-			{
-				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
-				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
-
-				ImGui.TableNextRow();
-
-				ImGui.TableNextColumn();
-				ImGui.Text("Entity Id");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.EditableEntityId(uniqueId, nameof(TransmuteEventData.EntityId), replay, ref e.EntityId);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt16Vec3(uniqueId, nameof(TransmuteEventData.A), ref e.A);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt16Vec3(uniqueId, nameof(TransmuteEventData.B), ref e.B);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt16Vec3(uniqueId, nameof(TransmuteEventData.C), ref e.C);
-
-				ImGui.TableNextColumn();
-				ImGui.Text("?");
-				ImGui.TableNextColumn();
-				EventTypeRendererUtils.InputInt16Vec3(uniqueId, nameof(TransmuteEventData.D), ref e.D);
-
-				ImGui.EndTable();
-			}
-
-			ImGui.SameLine();
-		}
-
-		ImGui.EndChild();
-	}
 }
