+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Timeline/EventTypes/EntityOrientation.cs
@@ -1,0 +1,41 @@
+using DevilDaggersInfo.Core.Replay.Events.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Events.EventTypes;
+using ImGuiNET;
+
+namespace DevilDaggersInfo.Tools.Ui.ReplayEditor.Timeline.EventTypes;
+
+public static class EntityOrientation
+{
+	public static void RenderEdit(int uniqueId, EntityOrientationEventData e, EditorReplayModel replay)
+	{
+		const float leftColumnWidth = 120;
+		const float rightColumnWidth = 160;
+		const float tableWidth = leftColumnWidth + rightColumnWidth;
+
+		if (ImGui.BeginChild(Inline.Span($"EntityOrientationEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
+		{
+			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
+			{
+				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
+				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
+
+				ImGui.TableNextRow();
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Entity Id");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.EditableEntityId(uniqueId, nameof(EntityOrientationEventData.EntityId), replay, ref e.EntityId);
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Orientation");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputInt16Mat3x3Square(uniqueId, nameof(EntityOrientationEventData.Orientation), ref e.Orientation);
+
+				ImGui.EndTable();
+			}
+		}
+
+		ImGui.EndChild();
+	}
+}
