+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Timeline/EventTypes/EntityPosition.cs
@@ -1,0 +1,41 @@
+using DevilDaggersInfo.Core.Replay.Events.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Events.EventTypes;
+using ImGuiNET;
+
+namespace DevilDaggersInfo.Tools.Ui.ReplayEditor.Timeline.EventTypes;
+
+public static class EntityPosition
+{
+	public static void RenderEdit(int uniqueId, EntityPositionEventData e, EditorReplayModel replay)
+	{
+		const float leftColumnWidth = 120;
+		const float rightColumnWidth = 160;
+		const float tableWidth = leftColumnWidth + rightColumnWidth;
+
+		if (ImGui.BeginChild(Inline.Span($"EntityPositionEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
+		{
+			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
+			{
+				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
+				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
+
+				ImGui.TableNextRow();
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Entity Id");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.EditableEntityId(uniqueId, nameof(EntityPositionEventData.EntityId), replay, ref e.EntityId);
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Position");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputInt16Vec3(uniqueId, nameof(EntityPositionEventData.Position), ref e.Position);
+
+				ImGui.EndTable();
+			}
+		}
+
+		ImGui.EndChild();
+	}
+}
