+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Timeline/EventTypes/PedeSpawn.cs
@@ -1,0 +1,71 @@
+using DevilDaggersInfo.Core.Replay.Events.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Events.EventTypes;
+using DevilDaggersInfo.Tools.Utils;
+using ImGuiNET;
+
+namespace DevilDaggersInfo.Tools.Ui.ReplayEditor.Timeline.EventTypes;
+
+public static class PedeSpawn
+{
+	private static readonly string[] _pedeTypeNamesArray = EnumUtils.PedeTypeNames.Values.ToArray();
+
+	public static void RenderEdit(int uniqueId, PedeSpawnEventData e, EditorReplayModel replay)
+	{
+		const float leftColumnWidth = 120;
+		const float rightColumnWidth = 160;
+		const float tableWidth = leftColumnWidth + rightColumnWidth;
+
+		if (ImGui.BeginChild(Inline.Span($"PedeSpawnEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
+		{
+			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
+			{
+				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
+				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
+
+				ImGui.TableNextRow();
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Type");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputByteEnum(uniqueId, nameof(PedeSpawnEventData.PedeType), ref e.PedeType, EnumUtils.PedeTypes, _pedeTypeNamesArray);
+
+				ImGui.TableNextColumn();
+				ImGui.Text("?");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputInt(uniqueId, nameof(PedeSpawnEventData.A), ref e.A);
+
+				ImGui.TableNextColumn();
+				ImGui.Text("?");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputVector3(uniqueId, nameof(PedeSpawnEventData.B), ref e.B, "%.0f");
+
+				ImGui.EndTable();
+			}
+
+			ImGui.SameLine();
+
+			if (ImGui.BeginTable("Right", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
+			{
+				ImGui.TableSetupColumn("RightText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
+				ImGui.TableSetupColumn("RightInput", ImGuiTableColumnFlags.None, rightColumnWidth);
+
+				ImGui.TableNextRow();
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Position");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputVector3(uniqueId, nameof(PedeSpawnEventData.Position), ref e.Position, "%.2f");
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Orientation");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputMatrix3x3Square(uniqueId, nameof(PedeSpawnEventData.Orientation), ref e.Orientation, "%.2f");
+
+				ImGui.EndTable();
+			}
+		}
+
+		ImGui.EndChild();
+	}
+}
