+++ src/DevilDaggersInfo.Tools/Ui/ReplayEditor/Timeline/EventTypes/SpiderEggSpawn.cs
@@ -1,0 +1,46 @@
+using DevilDaggersInfo.Core.Replay.Events.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Data;
+using DevilDaggersInfo.Tools.Ui.ReplayEditor.Events.EventTypes;
+using ImGuiNET;
+
+namespace DevilDaggersInfo.Tools.Ui.ReplayEditor.Timeline.EventTypes;
+
+public static class SpiderEggSpawn
+{
+	public static void RenderEdit(int uniqueId, SpiderEggSpawnEventData e, EditorReplayModel replay)
+	{
+		const float leftColumnWidth = 120;
+		const float rightColumnWidth = 160;
+		const float tableWidth = leftColumnWidth + rightColumnWidth;
+
+		if (ImGui.BeginChild(Inline.Span($"SpiderEggSpawnEdit{uniqueId}"), default, ImGuiChildFlags.AutoResizeY))
+		{
+			if (ImGui.BeginTable("Left", 2, ImGuiTableFlags.None, new(tableWidth, 0)))
+			{
+				ImGui.TableSetupColumn("LeftText", ImGuiTableColumnFlags.WidthFixed, leftColumnWidth);
+				ImGui.TableSetupColumn("LeftInput", ImGuiTableColumnFlags.None, rightColumnWidth);
+
+				ImGui.TableNextRow();
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Spawner Entity Id");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.EditableEntityId(uniqueId, nameof(SpiderEggSpawnEventData.SpawnerEntityId), replay, ref e.SpawnerEntityId);
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Position");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputVector3(uniqueId, nameof(SpiderEggSpawnEventData.Position), ref e.Position, "%.2f");
+
+				ImGui.TableNextColumn();
+				ImGui.Text("Target Position");
+				ImGui.TableNextColumn();
+				EventTypeRendererUtils.InputVector3(uniqueId, nameof(SpiderEggSpawnEventData.TargetPosition), ref e.TargetPosition, "%.2f");
+
+				ImGui.EndTable();
+			}
+		}
+
+		ImGui.EndChild();
+	}
+}
